requirements.txt
python-docx>=0.8.11
Pillow>=9.0.0
requests>=2.31.0
pyyaml>=6.0.1
argparse>=1.4.0
PyQt6>=6.4.0
PyQt6-Qt6>=6.4.0
PyQt6-sip>=13.4.0
openai>=1.3.0
python-dotenv>=1.0.0